refactor: put peer into separate file
