+++ src/ips.rs
@@ -16,11 +16,10 @@ use std::{
     str::FromStr,
 };
 
-use serde::{Deserialize, Serialize};
-
 use crate::{
     config::{GeoLocationMode, IPSConfiguration},
     graph_utils::{construct_graph, find_bridges},
+    peer::Peer,
     CrunchyState, Node,
 };
 
@@ -41,15 +40,6 @@ struct IpsState {
     pub betweenness_factors: NormalizationFactors,
     pub closeness_factors: NormalizationFactors,
     pub eigenvector_factors: NormalizationFactors,
-}
-
-/// Peer list structure containing peer list for each node
-#[derive(Clone, Serialize, Deserialize)]
-pub struct Peer {
-    /// IP address of the node
-    pub ip: IpAddr,
-    /// List of peers for the node
-    pub list: Vec<IpAddr>,
 }
 
 /// Internal structure for storing peer information
@@ -253,37 +243,6 @@ impl Ips {
 
     // Helper functions
 
-    /// Generate peerlist for given nodes based on their connections
-    fn generate_peerlist(&self, nodes: &[Node]) -> Vec<Peer> {
-        let mut peer_list = Vec::with_capacity(nodes.len());
-
-        for node in nodes {
-            peer_list.push(self.generate_peerlist_for_node(node, nodes));
-        }
-
-        peer_list
-    }
-
-    /// Generate peerlist for given node based on its connections
-    fn generate_peerlist_for_node(&self, node: &Node, nodes: &[Node]) -> Peer {
-        let mut peer_list_entry = Peer {
-            ip: IpAddr::from_str(node.ip.as_str()).expect(ERR_PARSE_IP),
-            list: Vec::with_capacity(node.connections.len()),
-        };
-
-        for peer in &node.connections {
-            if *peer >= nodes.len() {
-                continue;
-            }
-
-            peer_list_entry
-                .list
-                .push(IpAddr::from_str(nodes[*peer].ip.as_str()).expect(ERR_PARSE_IP));
-        }
-
-        peer_list_entry
-    }
-
     /// Generate state for IPS
     fn generate_state(&self, nodes: &[Node]) -> IpsState {
         let mut ips_state = IpsState {
@@ -323,7 +282,7 @@ impl Ips {
         let closeness = &nodes.iter().map(|n| n.closeness).collect::<Vec<f64>>();
         ips_state.closeness_factors = NormalizationFactors::determine(closeness);
 
-        ips_state.peer_list = self.generate_peerlist(nodes);
+        ips_state.peer_list = Peer::generate_peerlist(nodes);
 
         ips_state
     }
@@ -507,39 +466,6 @@ mod tests {
     use spectre::{edge::Edge, graph::Graph};
 
     use super::*;
-
-    #[test]
-    fn generate_peerlist_for_node_test() {
-        let ips_config = IPSConfiguration::default();
-        let ips = Ips::new(ips_config);
-
-        let nodes = vec![
-            Node {
-                ip: "0.0.0.0".to_string(),
-                connections: vec![1, 2],
-                ..Default::default()
-            },
-            Node {
-                ip: "1.0.0.0".to_string(),
-                connections: vec![0, 2],
-                ..Default::default()
-            },
-            Node {
-                ip: "2.0.0.0".to_string(),
-                connections: vec![0, 1],
-                ..Default::default()
-            },
-        ];
-
-        let peer = ips.generate_peerlist_for_node(nodes.get(0).unwrap(), &nodes);
-        assert_eq!(peer.list.len(), 2);
-        assert!(peer
-            .list
-            .contains(&IpAddr::from_str(&nodes.get(1).unwrap().ip).unwrap()));
-        assert!(peer
-            .list
-            .contains(&IpAddr::from_str(&nodes.get(2).unwrap().ip).unwrap()));
-    }
 
     #[test]
     fn rate_node_test() {
